Split data entry integration tests into their own file
